Address issues
More tricky test of DaemonStatus
Lock/Unlock are not longer exposed from CallbackDebounce
Moved Debounce to membership package so we can be more opinionated in
decisions
