--- common/debounce/channel.go
@@ -1,55 +1,0 @@
-// The MIT License (MIT)
-
-// Copyright (c) 2017-2020 Uber Technologies Inc.
-
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-package debounce
-
-import (
-	"time"
-
-	"github.com/uber/cadence/common/clock"
-)
-
-// DebouncedChannel is a wrapper around DebouncedCallback
-// providing channel instead of callback
-type DebouncedChannel struct {
-	*DebouncedCallback
-	signalCh chan struct{}
-}
-
-// NewDebouncedChannel creates a new NewDebouncedChannel which will
-// accumulate calls to .Handler function and will write data to Chan() on given interval if there were any calls to
-// .Handler in between. It is almost like rate limiter, but it never ignores calls to handler - only postpone.
-func NewDebouncedChannel(timeSource clock.TimeSource, interval time.Duration) *DebouncedChannel {
-	res := &DebouncedChannel{}
-
-	res.signalCh = make(chan struct{}, 1)
-	callback := func() {
-		select {
-		case res.signalCh <- struct{}{}:
-		default:
-		}
-	}
-	res.DebouncedCallback = NewDebouncedCallback(timeSource, interval, callback)
-	return res
-}
-
-func (d *DebouncedChannel) Chan() <-chan struct{} { return d.signalCh }
