+++ common/membership/debounce_test.go
@@ -20,9 +20,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package debounce
+package membership
 
 import (
+	"context"
+	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
@@ -31,6 +33,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.uber.org/goleak"
 
+	"github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/clock"
 )
 
@@ -38,6 +41,7 @@ const (
 	testDebounceInterval = time.Second
 	// this one is required since we're often testing for absence of extra calls, so we have to wait
 	testSleepAmount = time.Second / 2
+	testTimeout     = 10 * time.Second
 )
 
 type callbackTestData struct {
@@ -46,16 +50,9 @@ type callbackTestData struct {
 	calls             atomic.Int32
 }
 
-func waitCondition(fn func() bool, duration time.Duration) bool {
-	started := time.Now()
-
-	for time.Since(started) < duration {
-		if fn() {
-			return true
-		}
-		time.Sleep(time.Millisecond)
-	}
-	return false
+type channelTestData struct {
+	mockedTimeSource clock.MockedTimeSource
+	debouncedChannel *DebouncedChannel
 }
 
 func newCallbackTestData(t *testing.T) *callbackTestData {
@@ -78,6 +75,57 @@ func newCallbackTestData(t *testing.T) *callbackTestData {
 	td.mockedTimeSource.BlockUntil(1) // we should wait until ticker is created
 
 	return &td
+}
+
+func newChannelTestData(t *testing.T) *channelTestData {
+	var td channelTestData
+
+	td.mockedTimeSource = clock.NewMockedTimeSourceAt(time.Now())
+	td.debouncedChannel = NewDebouncedChannel(td.mockedTimeSource, testDebounceInterval)
+
+	t.Cleanup(
+		func() {
+			td.debouncedChannel.Stop()
+			goleak.VerifyNone(t)
+		},
+	)
+
+	td.debouncedChannel.Start()
+	// we should wait until ticker is created
+	td.mockedTimeSource.BlockUntil(1)
+	return &td
+}
+
+func (td *channelTestData) countCalls(t *testing.T, calls *atomic.Int32) {
+	// create a reader from channel which will save calls:
+	//   this way it should be easier to write tests by just evaluating `calls`
+	var wg sync.WaitGroup
+	ctx, cancel := context.WithCancel(context.Background())
+
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+
+		for {
+			select {
+			case <-td.debouncedChannel.Chan():
+				calls.Add(1)
+			case <-ctx.Done():
+				return
+			}
+		}
+	}()
+
+	t.Cleanup(
+		func() {
+			cancel()
+			assert.True(
+				t,
+				common.AwaitWaitGroup(&wg, testTimeout),
+				"debouncedChannel channel reader must be stopped",
+			)
+		},
+	)
 }
 
 func TestDebouncedCallbackWorks(t *testing.T) {
@@ -123,4 +171,74 @@ func TestDebouncedCallbackDoubleStopIsOK(t *testing.T) {
 
 	td.debouncedCallback.Stop()
 	assert.NotPanics(t, func() { td.debouncedCallback.Stop() }, "double stop should be OK")
+}
+
+func TestDebouncedSignalWorks(t *testing.T) {
+	td := newChannelTestData(t)
+
+	var calls atomic.Int32
+	td.countCalls(t, &calls)
+
+	td.debouncedChannel.Handler()
+	require.True(
+		t,
+		waitCondition(func() bool { return calls.Load() > 0 }, testTimeout),
+		"first callback is expected to be issued immediately after handler",
+	)
+	assert.Equal(t, 1, int(calls.Load()), 1)
+
+	// we call handler multiple times. There should be just one message in channel
+	for i := 0; i < 10; i++ {
+		td.debouncedChannel.Handler()
+	}
+
+	td.mockedTimeSource.Advance(testDebounceInterval)
+	time.Sleep(testSleepAmount)
+	assert.Equal(t, 2, int(calls.Load()))
+
+	// now call handler again, but advance time only by little - no messages in channel are expected
+	for i := 0; i < 10; i++ {
+		td.debouncedChannel.Handler()
+	}
+
+	td.mockedTimeSource.Advance(testDebounceInterval / 2)
+	time.Sleep(testSleepAmount)
+	assert.Equal(t, 2, int(calls.Load()), "should not have new messages in channel")
+}
+
+func TestDebouncedSignalDoesntDuplicateIfWeDontReadChannel(t *testing.T) {
+	td := newChannelTestData(t)
+
+	// this should lead message in channel
+	for i := 0; i < 10; i++ {
+		td.debouncedChannel.Handler()
+	}
+	td.mockedTimeSource.Advance(testDebounceInterval)
+	time.Sleep(testSleepAmount)
+
+	// same, but we don't expect new message since we haven't read one
+	for i := 0; i < 10; i++ {
+		td.debouncedChannel.Handler()
+	}
+	td.mockedTimeSource.Advance(testDebounceInterval)
+	time.Sleep(testSleepAmount)
+
+	// only now we start reading messages from channel
+	var calls atomic.Int32
+	td.countCalls(t, &calls)
+	time.Sleep(testSleepAmount)
+
+	assert.Equal(t, 1, int(calls.Load()), "Only a single message is expected")
+}
+
+func waitCondition(fn func() bool, duration time.Duration) bool {
+	started := time.Now()
+
+	for time.Since(started) < duration {
+		if fn() {
+			return true
+		}
+		time.Sleep(time.Millisecond)
+	}
+	return false
 }
